Script/Spells: removed useless cast made in previous commits

+++ src/world/Server/Script/ScriptMgr.cpp
@@ -92,7 +92,7 @@ SpellScriptExecuteState ScriptMgr::callScriptedSpellBeforeSpellEffect(Spell* spe
     if (!spell->getSpellInfo()->spellScript)
         return SpellScriptExecuteState::EXECUTE_NOT_HANDLED;
 
-    return spell->getSpellInfo()->spellScript->beforeSpellEffect(spell, effectType, static_cast<uint8_t>(effectId));
+    return spell->getSpellInfo()->spellScript->beforeSpellEffect(spell, effectType, effectId);
 }
 
 void ScriptMgr::callScriptedSpellAfterSpellEffect(Spell* spell, uint32_t effectType, uint8_t effectId)
@@ -100,7 +100,7 @@ void ScriptMgr::callScriptedSpellAfterSpellEffect(Spell* spell, uint32_t effectT
     if (!spell->getSpellInfo()->spellScript)
         return;
 
-    spell->getSpellInfo()->spellScript->afterSpellEffect(spell, effectType, static_cast<uint8_t>(effectId));
+    spell->getSpellInfo()->spellScript->afterSpellEffect(spell, effectType, effectId);
 }
 
 void ScriptMgr::register_spell_script(uint32_t spellId, SpellScript* ss)
